Switch time zone tests to default to local environment
In order to get the tests to pass on http://travis-ci.org, I altered them to work with UTC as the default. This required using `ENV TIMEZONE='American/New_York'` everytime I wanted to run the tests locally.

This change switches it so that the tests can be run locally the normal way, with the environment setting behind controlled through the Travis CI configuration file.