Add assertions for datetime objects
The `_add_time()`, `format()`, and `shift()` methods all required that the `value` argument contain a valid datetime object. They all now include `assert` statements to make sure this is the case.

In each instance, `assert` is provided with the result from the `_is_date_type()` method. This method checks that `value` is either a `datetime.datetime`, `datetime.date`, or `datetime.time`.