Add ever() method
The `ever()` method can be used to generate a random datetime object. This is useful for conducting tests that need a datetime, but don't require it to be anything specific.

Instead of using `datetime.MINYEAR` and `datetime.MAXYEAR` as the bounds, the current year +/- 100 are used. This is because if the date is too extreme (e.g., year 1), it will probably be less meaningful for tests.