Make when.formats compatible with Python 3
The way metaclasses work in Python 2 and 3 are different. The `__metaclass__` originally defined as part of `formats` has been moved to a separate class, `_FormatsMetaClass`. `formats` is now created dynamically. This method works for both versions of Python.

Thanks to Michael Watkins (@confute) for providing his awesome example at http://mikewatkins.ca/2008/11/29/python-2-and-3-metaclasses/#using-the-metaclass-in-python-2-x-and-3-x

Also changed 'pre-defined' to 'predefined' to be consistent.